requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.2
django-crispy-forms==1.11.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
folium==0.12.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.1.0
numpy==1.20.3
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
